+++ src/File.php
@@ -23,6 +23,10 @@ declare(strict_types=1);
 namespace Inane\File;
 
 use SplFileInfo;
+use Inane\Stdlib\{
+	String\Capitalisation,
+	Options
+};
 
 use function array_map;
 use function array_pop;
@@ -31,6 +35,7 @@ use function count;
 use function file;
 use function file_exists;
 use function file_get_contents;
+use function file_put_contents;
 use function floor;
 use function getcwd;
 use function glob;
@@ -48,17 +53,13 @@ use function strtolower;
 use function strtoupper;
 use function unlink;
 use function unserialize;
+
 use const DIRECTORY_SEPARATOR;
-use const false;
 use const FILE_APPEND;
 use const FILE_IGNORE_NEW_LINES;
 use const FILE_SKIP_EMPTY_LINES;
+use const LOCK_EX;
 use const null;
-
-use Inane\Stdlib\{
-	String\Capitalisation,
-	Options
-};
 
 /**
  * File metadata
@@ -298,10 +299,10 @@ class File extends SplFileInfo implements FSOInterface {
 	 *
 	 * @param string $contents to write to file
 	 *
-	 * @return bool success
+	 * @return bool|int false on failure otherwise number of bytes written.
 	 */
-	public function write(string $contents, bool $append = false, bool $createPath = true): bool {
-		$flag = $append ? FILE_APPEND : 0;
+	public function write(string $contents, bool $append = false, bool $createPath = true): bool|int {
+		$flag = $append ? FILE_APPEND | LOCK_EX : 0;
 
 		if (!$this->isValid())
 			$this->getParent()->makePath(recursive: $createPath);
@@ -316,7 +317,7 @@ class File extends SplFileInfo implements FSOInterface {
 
 			$this->contentCache->array = null;
 
-			return true;
+			return $success;
 		}
 
 		return false;
@@ -331,9 +332,9 @@ class File extends SplFileInfo implements FSOInterface {
 	 *
 	 * @param string $contents to append to file
 	 *
-	 * @return bool success
+	 * @return bool|int false on failure otherwise number of bytes written.
 	 */
-	public function append(string $contents): bool {
+	public function append(string $contents): bool|int {
 		return $this->write($contents, true);
 	}
 
@@ -344,9 +345,9 @@ class File extends SplFileInfo implements FSOInterface {
 	 *
 	 * @param string $contents to write to file
 	 *
-	 * @return bool success
+	 * @return bool|int false on failure otherwise number of bytes written.
 	 */
-	public function prepend(string $contents): bool {
+	public function prepend(string $contents): bool|int {
 		$contents .= $this->read();
 		return $this->write($contents);
 	}